Make a nanodrop object

Make a nanodrop object

Usage

incucyte(data, platemap)

Arguments

data

a data.frame - or something coercible to one - containing nanodrop data. See details for requirements.

date

an optional character vector or something coercible to a Date object with lubridate::as_date

nucleotide

a character vector to specifying the analyte in the nanodrop experiment. Must match one of the following: RNA, DNA, dsDNA, or dsRNA.

is_tidy

logical. Is the supplied data tidy?

Value

an incucyte object

Details

Making a nanodrop object 'by hand' (that is, not using read_nanodrop) is not recommended, since it is challenging to ensure a given data.frame is truly a valid nanodrop file. To increase reliability, flexibility is reduced. As such, the supplied data.frame must at least have the following columns if is_tidy = TRUE:

date

sample_name

conc

a260_280

a230_280

If is_tidy = FALSE, at least the following columns must be provided:

Date

Sample.Name

Nucleic.Acid.ng.uL.

A260.A280

A260.A230

Note: technically, the given column names will be stripped of all non-alphanumerics and forced tolower, then compared against the following:

date

samplename

nucleicacidngul

a260a280

a260a230

In both cases, other columns allowed are those that appear in Example A and Example B below.

If there are additional columns provided, they will be silently dropped.