Constructor for a pcr object

Constructor for a pcr object

Usage

new_pcr(
  data = data.frame(),
  raw_data = raw(),
  header = character(),
  footer = character(),
  date = lubridate::Date(),
  experiment_type = character(),
  wells = integer(),
  is_tidy = logical()
)

Arguments

data

data.frame or gp, no restrictions on form.

raw_data

raw, read in from the provided file with readr::read_file_raw()

header

data.frame, the upper portion of the data before the body of the data. Usually contains run information.

footer

data.frame, the lower portion of the data after the body of the data. Usually contains analysis information. May not exist.

date

lubridate::Date object

experiment_type

character. Could

wells

Number of wells in plate

is_tidy

logical. Is the data.frame provided tidy?

Value

A pcr object