Tidy a nanodrop object

Tidy a nanodrop object

Usage

# S3 method for class 'nanodrop'
tidy_lab(x, force_tidy = FALSE, ...)

Arguments

x

A nanodrop object to tidy

force_tidy

logical. Should the tidying take place, even if the is_tidy attribute is TRUE?

...

Unused

Value

a nanodrop object

Details

This function:

renames columns to sensible substitutes

converts date-times to ISO 8601-esque date-time format (YYYY-MM-DD HH:MM:SS vs YYYY-MM-DDTHH:MM:SSZ)

It's recommended that you do any manipulations to these data after you tidy, rather than before, as tidy_nanodrop expects the output to be fairly similar to the output from read_nanodrop.

A tidy nanodrop object will (usually) contain the following columns:

date The date-time of sample reading, as YYYY-MM-DD HH:MM:SS

sample_name The name of the sample provided by the NanoDrop itself

conc The concentration of nucleic acid in ng/uL

a260_280 The absorbance at 260nm / absorbance at 280nm. Typically a marker of protein contamination. Pure nucleic acid is typically around 2.

a260_230 The absorbance at 260nm / absorbance at 230nm. Typically a marker of guanadine salt contamination. Pure nucleic acid is typically around 2.

a260 The absorbance at 260nm, the wavelength nucleic acids absorb most strongly.

a280 The absorbance at 280nm, the wavelength proteins absorb most strongly.

The remaining columns are typically unused.